genvm/requirements.txt
Flask-JSONRPC==3.0.1
Flask[async]==3.0.2
aiohttp==3.9.3
python-dotenv==1.0.1